Remove explicit minOccurs on top-level elements
A top-level xs:sequence or xs:choice should have all its contained items
have minOccurs=0, so there is no need to have minOccurs=0 on the parent.
This provides visual consistency, particularly with attributes, which
show a darker grey line.

The FragmentType is still a little strange. It has an explicit
minOccurs=0, maxOccurs=umbounded. Perhaps this should be set on the
parent type instead.