Change datatypes on BaseObjectProperty subtypes
Change them from being "fixed" to "default". This allows producers to
encode a property in a different type than what is proscribed for that
type. For example:

- if a "string" cannot be recorded as UTF-8 string, use Base64Binary
instead
- if you want to record integer fields using their hexadecimal
equivalent.

I did not change BaseObjectProperty subtypes where a union of some enum
and "string" was used, since these are always meant to be strings, thus
"fixed" is appropriate.

Fixes #3, Fixes #112